worker/requirements.txt
celery==5.4.0
redis==5.0.8
boto3==1.35.28
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
elasticsearch==8.15.1
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
scikit-learn==1.5.2
rapidfuzz==3.9.7
python-docx==1.1.2
PyMuPDF==1.24.10
pdfminer.six==20231228
python-dotenv==1.0.1
pydantic==2.9.2

# AI Detection
transformers==4.45.0
torch==2.5.0
